adjusting project structure;
not perfect yet. how can we decouple "app" from "views"? i'd like to
have all views in a dedicated subfolder/-module, but i'm not aware of
how i can then define methods on the App struct from within that
submodule. i can define a type alias, but would then require to import
the submodule from the "app" module, which will in turn trigger the
compiler to complain about an "unused import" and thus it will optimize
the import away, resulting in the views not being recognized.